Adds API for automation modules
Modules are listed inside lib/automate/modules. Modules are
required to have the attribute self.verbs and the methods
self.run(self, to, when, body, sender) as well as
self.followup(self, answer). An example module is shown in
lib/automate/modules/send.py, which sends a mail given the
parameters.

The when-parameter is currently ignored, but should
in the future be a datetime.datetime object which defines when the
command should be executed.

The example does not use the self.followup(self, answer). The
intended use of this is to ask extra questions if something in
the original call was unclear.