Add functonality for CLI to call automation module
With the CLI running the user can now enter a similar word to any
existing automation module and the module will be called with predefined
arguments. Since the arguments to the automation module are predefined
only the Send module will work correctly if we add more in the future.

For example entering `skicka` will correctly invoke the Send module and
the predefined email will be sent.

A generic user was also added to the settings.py file and if the sender
is specified as `"John Doe"` this user will be used in the automation
module. The email of this user is setup to work if you have a local SMTP
server running on `localhost:1025`, see the README for how to start one.

Updated the README to include instructions on how to setup a local SMTP
sever and some CLI usage info.