Add Reminder automation module
Implemented a reminder automation module which given a time and body
will schedule a reminder which will pop-up in the users OS. Currently
only Linux is supported and this will have to be extended to also work
on Windows 10 in the future. This can easily be done by adding code to
`Reminder.notify` which invokes something in the windows system that can
create a notification and adding `win32` to the `supported_os[]`.

Reminer.run will raise an error if the time at which the reminder is
supposed to happen has already passed or if the function is called from
a host OS that is not currently supported.

To test this you can run `reminder_demo.py`. The demo is currently not
using the NLP module since I had some problems with how the NLP parsed
time in the text, this will have to be changed in the future.

I also wrote some tests for the reminder module, in these tests I also
mocked some imported library functions so you can check how this can be
done using fixtures if you need to use it in your own tests. The fixtures
can be found in the `conftest.py` file.