Add util function that resolves emails
Created a new util function which is able to resolve the email addresses
of local contacts given their names (or valid emails directly). This is
done by returning a dictionary containing:
- emails: the known emails resolved
- uncertain: the name and potential candidates and their emails, will
contain any contacts which could not be resolved i.e. multiple with same
name
- unknown: the input names which could not be resolved to any local
contacts

This function is now used when trying to resolve the emails of meeting
attendees. Necessary follow-ups has also been added such as which
contanct to pick (if uncertain is not empty) and if the meeting should
still be scheduled without the unknown contact.