Add Speech-to-Text layer
This has been implemented using the Google Cloud Speech-to-Text API. To
do this a microphone stream had to be implemented. The Speech-to-Text
layer will transcribe one spoken sentence and return the final
interpreted text. This has also been added to the CLI and can be used by
issuing the `record` command.