Creates a followup object
Earlier the followup was handled by the automation modules
themselves, returing only a string for context.

This commit defines a Followup interface, with inherited classes
MultiFollowup, StringFollowup and BooleanFollowup. The attributes
of the different classes differ somewhat, but they all contain
a .question and an .answer. Control can be given back to the
automation module via a .callback().