Add missing settings and validation of fields (#179)
Added all of the settings required to use the program.

Also added validation to most of the settings fields, this includes
- Editor
  - font-size: Not empty and only numbers allowed
- Meetings
  - standard duration: Not empty and only numbers allowed
- User
  - name: Not empty
  - email: Not empty and only valid emails allowed
These fields will also be styled as invalid whenever the input is
invalid.

Also, the password field will not show the password (dots). The user is
also only allowed to select the model languages and versions
defined in `./config/nlp_models.yaml`.